feat: validators table by session query and stithc
