+++ front/gatsby/src/components/index.ts
@@ -15,3 +15,4 @@ export * from './Status';
 export * from './Table';
 export * from './Tooltip';
 export * from './Typography';
+export { default as ValidatorsTable } from './ValidatorsTable';
