+++ front/gatsby/src/util/graphql.ts
@@ -55,9 +55,10 @@ export const CURRENT_ELECTED = gql`
   }
 `;
 
+
 export const CURRENT_NOMINATIONS = gql`
   query Nominators($sessionIndex: Int!) {
-    nominators(where: { session: { index: $sessionIndex } }, last: 150) {
+    nominations(where: { session: { index: $sessionIndex }}) {
       validatorController
       validatorStash
       nominatorStash
@@ -69,11 +70,11 @@ export const CURRENT_NOMINATIONS = gql`
 
 export const OFFLINE_VALIDATORS = gql`
   query OfflineValidators($sessionIndex: Int!) {
-    offlineValidators(last: 40) {
+    offlineValidators(where: { session: { index: $sessionIndex }}) {
       validatorId
-      # total
-      # own
-      # others
+      total
+      own
+      others
     }
   }
 `;
